New Home page and menu updates
