+++ styles/globals.css
@@ -21,6 +21,10 @@ body {
   font-family: "Verdana", "sans-serif";
   font-size: var(--base-font-size);
 }
+h1,h2,h3,h4{
+  margin: 0;
+  padding:0.75rem 0;
+}
 
 input{
   font-size: var(--base-font-size);
